// tms_v7.9_beta_nzwcf0/convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as auth from "../auth.js";
import type * as carriers from "../carriers.js";
import type * as clients from "../clients.js";
import type * as dashboard from "../dashboard.js";
import type * as http from "../http.js";
import type * as loads from "../loads.js";
import type * as locations from "../locations.js";
import type * as router from "../router.js";
import type * as tasks from "../tasks.js";
import type * as users from "../users.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  carriers: typeof carriers;
  clients: typeof clients;
  dashboard: typeof dashboard;
  http: typeof http;
  loads: typeof loads;
  locations: typeof locations;
  router: typeof router;
  tasks: typeof tasks;
  users: typeof users;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;